==> 2-LastFM.ipynb <==
#### _______IMPORTS Y AUTENTICACIONES_______

# IMPORTS
import requests
import pandas as pd
import csv

# AUTENTICACIÓN
api_key = '3784874dc3142503331b4556466b8965'  
base_url = 'http://ws.audioscrobbler.com/2.0/'


### _________LECTURA CSV DESDE SPOTIPY Y CONVERSIÓN EN LISTA_________

lista_artistas = pd.read_csv("artistas_v2.csv")
nombres_artistas = lista_artistas["artist"].tolist() 

COMPROBACIONES

len(nombres_artistas)

# ________CONSULTA DE BIO, ARTISTAS SIMILARES, REPRODUCCIONES, TAGS Y TOUR________

artistasFM = []  # Lista donde guardaremos los diccionarios de cada artista

for artist in nombres_artistas:   # Queremos que para cada artista de la lista, recorra cada uno de ellos y obtengamos su información
    params_info = {               # Son los parámetros de la petición de la API
        'method': 'artist.getinfo',
        'artist': artist,
        'api_key': api_key,
        'format': 'json'} 
    try:                                 
        response_info = requests.get(base_url, params=params_info)  # Response nos indica si hay algún error en la petición
        response_info.raise_for_status()         # Si la respuesta tiene un código de error (tipo 400), se va al except
                                
        data_json = response_info.json()  # Extraemos los datos de la API en formato JSON
        nombre_artista = data_json.get('artist', {})  # Si no encuentra el dato, nos devuelve un diccionario vacío
        
        # Reproducciones y oyentes
        stats = nombre_artista.get('stats', {})       # stats es el sub-diccionario donde suele venir info como playcount y listeners 
        playcount = stats.get('playcount', 0)         # Total de reproducciones. Devolveremos 0 si la clave no existe
        listeners = stats.get('listeners', 0)         # Número de oyentes. Devolveremos 0 si la clave no existe
        
        # Biografía resumida
        biografia = nombre_artista.get('bio', {}).get('summary', '')  # Resumen de la biografía. Si no existe, devuelve cadena vacía
        
        # Artistas similares
        artistas_similares = []
        similar_artists = nombre_artista.get('similar', {}).get('artist', [])   # Seguimos el mismo patrón
        for s in similar_artists:
            artistas_similares.append(s['name'])
        
        # Tags del artista (como otros géneros asociados)
        tags_artist = []
        tags_list = nombre_artista.get('tags', {}).get('tag', [])
        for t in tags_list:
            tags_artist.append(t['name'])
        
        # Estado de tour
        ontour = nombre_artista.get('ontour', '0')  # 0 = no está de gira, 1 = sí
        
        # Creamos un diccionario con toda la información del artista
        artistasFM.append({      
            'Nombre Artista': artist,
            'Reproducciones': playcount,
            'Oyentes': listeners,
            'Biografia': biografia,
            'Artistas Similares': ", ".join(artistas_similares), # convierte la lista en una cadena separada por comas
            'Tags': ", ".join(tags_artist),
            'Ontour': ontour})      
        
    except requests.exceptions.RequestException as excep:   # Si ocurre un error de red o de la API, lo imprime y continúa
        print(f"El error para el artista {artist}: {excep}")


### _________CONVERSIÓN A DATAFRAME_________

df_artistas = pd.DataFrame(artistasFM)
df_artistas

### _________CONVERSIÓN A CSV_________

df_artistas.to_csv('bio_artistas_v2.csv',index=False)
