==> 3-Conexion_Python-SQL.ipynb <==
# ________CONEXIÓN PYTHON A SQL________

#### _______IMPORTS_______

import pandas as pd
import mysql.connector
from mysql.connector import errorcode
from sqlalchemy import create_engine,FLOAT, VARCHAR, INTEGER, DATE, SmallInteger
from sqlalchemy.sql.sqltypes import String

#### _______LECTURA DE FICHEROS_______

df_albumes = pd.read_csv("albums_v2.csv")
df_canciones = pd.read_csv("canciones_v2.csv")
df_bio= pd.read_csv("bio_artistas_v2.csv")

#### _______CONEXIÓN CON SQL_______

engine = create_engine(
    "mysql+mysqlconnector://root:AlumnaAdalab@127.0.0.1/stella_apis_bbdd",
    connect_args={
        "auth_plugin": "mysql_native_password"})


df_albumes.to_sql("albums_v2", con=engine, if_exists="replace", index=False)
df_canciones.to_sql("songs_v2", con=engine, if_exists="replace", index=False)
df_bio.to_sql("bio_artistas_v2", con=engine, if_exists="replace", index=False)
